graph_app/requirements.txt
grpcio==1.29.0
grpcio-tools==1.29.0
protobuf==3.12.2
pybuilder==0.12.6
flask==1.1.2
multipledispatch==0.6.0
abcplus==0.1.0
azure-storage-blob==12.3.2
azure-core==1.6.0
requests==2.23.0
python-dateutil==2.8.1
Flask-Cors